# gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package gen

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Error defines model for Error.
type Error struct {
	Message    *string `json:"message,omitempty"`
	StatusCode *int64  `json:"statusCode,omitempty"`
	Title      *string `json:"title,omitempty"`
}

// N400Error defines model for 400Error.
type N400Error = Error

// N401UnauthorizedError defines model for 401UnauthorizedError.
type N401UnauthorizedError = Error

// N500Error defines model for 500Error.
type N500Error = Error

// PutAccountJSONBody defines parameters for PutAccount.
type PutAccountJSONBody struct {
	FamilyName     string `json:"familyName"`
	FamilyNameKana string `json:"familyNameKana"`
	FirstName      string `json:"firstName"`
	FirstNameKana  string `json:"firstNameKana"`
}

// PutAccountJSONRequestBody defines body for PutAccount for application/json ContentType.
type PutAccountJSONRequestBody PutAccountJSONBody